backend-master/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
redis
minio
argon2
pyotp
Flask-Limiter 
pytest
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests

